Don't coalesce user find requests
This is to handle the case where a user enters the app through the
Friend Show route. Ember helpfully tries to coalesce the requests for
the current user and the friend into a single request, but since we're
treating the current user as a singular resource with an id of
'current', our endpoint gets 'current' in the list of requested ids.
It's easier to just not bother coalescing requests than it is to
manually handle the case when the current user is requested.

Maybe later we can revisit how we've implemented the current user
singular resource, but for now this is the easiest thing to do.